5.4 fixes

+++ src/main/java/mods/thecomputerizer/musictriggers/MusicTriggers.java
@@ -39,9 +39,6 @@ import java.util.List;
 public class MusicTriggers {
     public static final String MODID = "musictriggers";
 
-    public static File songsDir;
-    public static File texturesDir;
-
     private static RepositorySource source;
 
     public static final Logger logger = LogManager.getLogger();
@@ -53,34 +50,20 @@ public class MusicTriggers {
         IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
         MinecraftForge.EVENT_BUS.register(this);
         File configDir = new File("config", "MusicTriggers");
-        if (!configDir.exists()) {
-            configDir.mkdir();
-        }
+        if (!configDir.exists()) configDir.mkdir();
         if (FMLEnvironment.dist == Dist.CLIENT) {
-            songsDir = new File(configDir.getPath(), "songs");
-            if (!songsDir.exists()) {
-                songsDir.mkdir();
-            }
+            File songsDir = new File(configDir.getPath(), "songs");
+            if (!songsDir.exists()) songsDir.mkdir();
             File assetsDir = new File(songsDir.getPath(), "assets");
-            if (!assetsDir.exists()) {
-                assetsDir.mkdir();
-            }
+            if (!assetsDir.exists()) assetsDir.mkdir();
             File musictriggersDir = new File(assetsDir.getPath(), "musictriggers");
-            if (!musictriggersDir.exists()) {
-                musictriggersDir.mkdir();
-            }
+            if (!musictriggersDir.exists()) musictriggersDir.mkdir();
             File soundsDir = new File(musictriggersDir.getPath(), "sounds");
-            if (!soundsDir.exists()) {
-                soundsDir.mkdir();
-            }
+            if (!soundsDir.exists()) soundsDir.mkdir();
             File musicDir = new File(soundsDir.getPath(), "music");
-            if (!musicDir.exists()) {
-                musicDir.mkdir();
-            }
-            texturesDir = new File(musictriggersDir.getPath(), "textures");
-            if (!texturesDir.exists()) {
-                texturesDir.mkdir();
-            }
+            if (!musicDir.exists()) musicDir.mkdir();
+            File texturesDir = new File(musictriggersDir.getPath(), "textures");
+            if (!texturesDir.exists()) texturesDir.mkdir();
             File mcmeta = new File(songsDir.getPath() + "/pack.mcmeta");
             if (!mcmeta.exists()) {
                 try {
@@ -96,18 +79,16 @@ public class MusicTriggers {
                     ex.printStackTrace();
                 }
             }
-            File jjson = new File(musictriggersDir.getPath() + "/sounds.json");
-            if (!jjson.exists() && Json.allSongs != null) {
+            File sjson = new File(musictriggersDir.getPath() + "/sounds.json");
+            if (!sjson.exists() && Json.allSongs != null) {
                 try {
-                    jjson.createNewFile();
+                    sjson.createNewFile();
                 } catch (IOException ex) {
                     ex.printStackTrace();
                 }
             }
             File langDir = new File(musictriggersDir.getPath(), "lang");
-            if (!langDir.exists()) {
-                langDir.mkdir();
-            }
+            if (!langDir.exists()) langDir.mkdir();
             File lang = new File(langDir.getPath() + "/en_us.json");
             if (!lang.exists()) {
                 try {
@@ -120,14 +101,14 @@ public class MusicTriggers {
             makeSoundsJson();
             makeDiscLang();
         }
-            ConfigToml.parse(FMLEnvironment.dist == Dist.CLIENT);
+        ConfigToml.parse(FMLEnvironment.dist == Dist.CLIENT);
         if(FMLEnvironment.dist == Dist.CLIENT) {
             ConfigCommands.parse();
             ConfigTitleCards.parse();
             if(Json.allSongs!=null) {
                 File pack = new File("config/MusicTriggers/");
                 if (pack.isDirectory()) {
-                    MusicTriggers.logger.info("found pack :)");
+                    MusicTriggers.logger.info("Loading songs as a resource pack");
                     source = new PackFolder(pack, PackSource.DEFAULT);
                 }
             }
@@ -177,17 +158,13 @@ public class MusicTriggers {
 
     public static void makeSoundsJson() {
         File sj = new File("config/MusicTriggers/songs/assets/musictriggers/sounds.json");
-        if (sj.exists()) {
-            sj.delete();
-        }
+        if (sj.exists()) sj.delete();
         List<String> writeThis = Json.create();
         if (writeThis != null) {
             try {
                 sj.createNewFile();
                 FileWriter writer = new FileWriter(sj);
-                for (String str : writeThis) {
-                    writer.write(str + System.lineSeparator());
-                }
+                for (String str : writeThis) writer.write(str + System.lineSeparator());
                 writer.close();
             } catch (IOException ex) {
                 ex.printStackTrace();
@@ -198,9 +175,7 @@ public class MusicTriggers {
     public static void makeDiscLang() {
         if(ConfigRegistry.registerDiscs) {
             File sj = new File("config/MusicTriggers/songs/assets/musictriggers/lang/en_us.json");
-            if (sj.exists()) {
-                sj.delete();
-            }
+            if (sj.exists()) sj.delete();
             List<String> writeThis = Json.create();
             assert writeThis != null;
             writeThis.clear();
@@ -209,9 +184,7 @@ public class MusicTriggers {
                 try {
                     sj.createNewFile();
                     FileWriter writer = new FileWriter(sj);
-                    for (String str : writeThis) {
-                        writer.write(str + System.lineSeparator());
-                    }
+                    for (String str : writeThis) writer.write(str + System.lineSeparator());
                     writer.close();
                 } catch (IOException ex) {
                     ex.printStackTrace();
